Remove coherence
