--- examples/integrations/coherence/pom.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright (c) 2021, 2024 Oracle and/or its affiliates.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
-                             https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.helidon.examples.integrations</groupId>
-        <artifactId>helidon-examples-integrations-project</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-    </parent>
-    <groupId>io.helidon.examples.integrations.coherence</groupId>
-    <artifactId>helidon-examples-integrations-coherence-project</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-    <packaging>pom</packaging>
-    <name>Helidon Examples Integration Coherence</name>
-    <description>Examples of integration with Coherence CE.</description>
-
-    <modules>
-        <module>helidon-creditscore-mp</module>
-    </modules>
-</project>
