+++ examples/microprofile/coherence/README.md
