+++ examples/microprofile/coherence/src/main/java/io/helidon/examples/microprofile/coherence/CreditscoreService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.examples.creditscore;
+package io.helidon.examples.microprofile.coherence;
 
 import com.oracle.coherence.cdi.Name;
 import com.tangosol.net.NamedCache;
