+++ examples/microprofile/coherence/src/main/java/io/helidon/examples/microprofile/coherence/Person.java
@@ -13,5 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.helidon.examples.microprofile.coherence;
 
-package io.helidon.examples.creditscore;
+public record Person(
+    String firstName,
+    String lastName,
+    String dateOfBirth,
+    String ssn,
+    int creditScore) {}
