+++ examples/microprofile/coherence/src/main/resources/META-INF/beans.xml
