+++ examples/microprofile/coherence/src/main/resources/META-INF/microprofile-config.properties
