+++ examples/microprofile/coherence/src/main/resources/logging.properties
