+++ examples/microprofile/coherence/src/test/resources/META-INF/microprofile-config.properties
